Extract AppHost configuration and add AppHost tests
