+++ Hagalaz.sln
@@ -37,6 +37,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hagalaz.Database.Migrations
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hagalaz.Database.Migrations.Tests", "Hagalaz.Database.Migrations.Tests\Hagalaz.Database.Migrations.Tests.csproj", "{D46576A8-A7A1-4A9D-B5C4-BCB5BB8DC73F}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hagalaz.AppHost.Tests", "Hagalaz.AppHost.Tests\Hagalaz.AppHost.Tests.csproj", "{3C6C9767-6BE9-4D66-BF94-CE369058B7BF}"
+EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hagalaz.Services.Authorization", "Hagalaz.Services.Authorization\Hagalaz.Services.Authorization.csproj", "{89B9AA78-69E2-40E9-A0D8-7E9A0FAAF741}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Hagalaz.Game.Abstractions", "Hagalaz.Game.Abstractions\Hagalaz.Game.Abstractions.csproj", "{9C9FA52B-FC05-4DF5-8E7E-5AFD32A70D6B}"
@@ -357,6 +359,18 @@ Global
 		{D46576A8-A7A1-4A9D-B5C4-BCB5BB8DC73F}.Release|x64.Build.0 = Release|Any CPU
 		{D46576A8-A7A1-4A9D-B5C4-BCB5BB8DC73F}.Release|x86.ActiveCfg = Release|Any CPU
 		{D46576A8-A7A1-4A9D-B5C4-BCB5BB8DC73F}.Release|x86.Build.0 = Release|Any CPU
+		{3C6C9767-6BE9-4D66-BF94-CE369058B7BF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3C6C9767-6BE9-4D66-BF94-CE369058B7BF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3C6C9767-6BE9-4D66-BF94-CE369058B7BF}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{3C6C9767-6BE9-4D66-BF94-CE369058B7BF}.Debug|x64.Build.0 = Debug|Any CPU
+		{3C6C9767-6BE9-4D66-BF94-CE369058B7BF}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{3C6C9767-6BE9-4D66-BF94-CE369058B7BF}.Debug|x86.Build.0 = Debug|Any CPU
+		{3C6C9767-6BE9-4D66-BF94-CE369058B7BF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3C6C9767-6BE9-4D66-BF94-CE369058B7BF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3C6C9767-6BE9-4D66-BF94-CE369058B7BF}.Release|x64.ActiveCfg = Release|Any CPU
+		{3C6C9767-6BE9-4D66-BF94-CE369058B7BF}.Release|x64.Build.0 = Release|Any CPU
+		{3C6C9767-6BE9-4D66-BF94-CE369058B7BF}.Release|x86.ActiveCfg = Release|Any CPU
+		{3C6C9767-6BE9-4D66-BF94-CE369058B7BF}.Release|x86.Build.0 = Release|Any CPU
 		{89B9AA78-69E2-40E9-A0D8-7E9A0FAAF741}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{89B9AA78-69E2-40E9-A0D8-7E9A0FAAF741}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{89B9AA78-69E2-40E9-A0D8-7E9A0FAAF741}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -1109,6 +1123,7 @@ Global
 		{9C3D2FA2-1658-4C88-95C9-744B77A47261} = {EDA310BD-1441-40DA-8514-81E6EEC9F9B3}
 		{89CEFFCB-CE29-4C2F-B108-C908EEA15EDC} = {EDA310BD-1441-40DA-8514-81E6EEC9F9B3}
 		{D46576A8-A7A1-4A9D-B5C4-BCB5BB8DC73F} = {EDA310BD-1441-40DA-8514-81E6EEC9F9B3}
+		{3C6C9767-6BE9-4D66-BF94-CE369058B7BF} = {EDA310BD-1441-40DA-8514-81E6EEC9F9B3}
 		{89B9AA78-69E2-40E9-A0D8-7E9A0FAAF741} = {EDA310BD-1441-40DA-8514-81E6EEC9F9B3}
 		{9C9FA52B-FC05-4DF5-8E7E-5AFD32A70D6B} = {76B390E4-B127-478D-946F-E44081AD9D6A}
 		{72F89E88-CB9E-4993-B5C1-0BA14E72F48A} = {76B390E4-B127-478D-946F-E44081AD9D6A}
